+++ packages/expo-modules-core/android/src/main/cpp/JavaCallback.h
@@ -53,7 +53,6 @@ private:
 
   friend HybridBase;
 
-
   JavaCallback(std::shared_ptr<CallbackContext> callback);
 
   void invoke();