+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/jni/JSIContext.kt
@@ -120,11 +120,6 @@ class JSIContext : Destructible {
    */
   external fun drainJSEventLoop()
 
-  /**
-   * Informs C++ that runtime was deallocated.
-   */
-  external fun wasDeallocated()
-
   external fun setNativeStateForSharedObject(id: Int, js: JavaScriptObject)
 
   /**